Import function (V1.0)
A function for importing players of a team given the team name.
Numbers 0 through 15 are assigned to the teams provided in the ZIP file, in that order. [0 assigned to Esteghlal and so forth]
Also, a global variable of type "Team" was needed in order for this function to perform properly.